requirements.txt
mss
numpy
opencv-python
pillow
pytesseract